+++ packages/paste-core/utilities/grid/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.1](https://github.com/twilio-labs/paste/compare/@twilio-paste/grid@0.1.0...@twilio-paste/grid@0.1.1) (2020-01-30)
+
+
+### Bug Fixes
+
+* **grid:** set useTheme as const to fix hook error ([#284](https://github.com/twilio-labs/paste/issues/284)) ([851a188](https://github.com/twilio-labs/paste/commit/851a188c7bc9a50e50b63500ec3ff5782755339a))
+
+
+
+
+
+# [0.1.0](https://github.com/twilio-labs/paste/compare/@twilio-paste/grid@0.0.2...@twilio-paste/grid@0.1.0) (2020-01-30)
+
+
+### Features
+
+* **grid:** create grid component ([f43f82a](https://github.com/twilio-labs/paste/commit/f43f82af6362afcb5d0c0d85cfc2560d441e3ff9))
+
+
+
+
+
 ## [0.0.2](https://github.com/twilio-labs/paste/compare/@twilio-paste/grid@0.0.1...@twilio-paste/grid@0.0.2) (2019-10-29)
 
 **Note:** Version bump only for package @twilio-paste/grid